# public/swagger.yaml
openapi: 3.0.0
info:
  description: 'Access token should be sent along with every request to SITE_NAME API in authorization header: `Authorization: Bearer <Token>` <br> This token can either be acquired via `auth/login` endpoint or from account settings page on [SITE_NAME website](SITE_URL/account-settings).'
  version: '1.0.0'
  title: SITE_NAME API
security:
  - accessToken: []
...
paths:
...
  /file-entries/{entryId}/shareable-link:
    get:
      tags:
        - Shareable Links
      summary: Retrieve shareable link for specified entry
      operationId: getShareableLink
      parameters:
        - name: entryId
          in: path
          description: ID or hash of entry for which to get shareable link
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shareable-Link-Response'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
        '404':
          description: File entry or shareable link not found
        '422':
          description: Invalid data specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/422-Response'
    post:
      tags:
        - Shareable Links
      summary: Create new shareable link for specified entry
      operationId: createShareableLink
      parameters:
        - name: entryId
          in: path
          description: ID or hash of entry for which to get shareable link
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: 'new password'
                expires_at:
                  type: string
                  description: '2021-03-06T17:34:00.000000Z'
                allow_edit:
                  type: boolean
                  example: false
                allow_download:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Shareable link created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  link:
                    $ref: '#/components/schemas/ShareableLink'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
        '422':
          description: Invalid data specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/422-Response'
...
components:
  schemas:
    FileEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: image.png
        file_name:
          type: string
          example: j6oWHDG1z613UqFHLZpE1Yrqngw0N8XLuGNLNWJV
          description: actual file name for this entry on storage
        file_size:
          type: integer
          format: int64
          example: 111863
          description: In bytes
        parent_id:
          type: integer
          format: int64
          example: 1
          description: Parent folder ID
        parent:
          $ref: '#/components/schemas/FileEntry'
        thumbnail:
          type: string
          example: thumbnail.png
          description: Relative path to thumbnail image for the file (if it exists)
        mime:
          type: string
          example: image/png
        url:
          type: string
          example: secure/uploads/3260
          description: relative url for previewing file or folder contents
        hash:
          type: string
          example: MzI2MHxwYWRkaQ
        type:
          type: string
          enum:
            - 'image'
            - 'folder'
            - 'text'
            - 'audio'
            - 'video'
            - 'pdf'
        description:
          type: string
          example: Example description entered by user
        deleted_at:
          type: string
          example: '2021-02-23T14:42:38.000000Z'
          description: date entry was moved to trash (if it's in trash currently)
        created_at:
          type: string
          example: '2021-02-23T14:42:38.000000Z'
          description: date entry was uploaded
        updated_at:
          type: string
          example: '2021-02-23T14:42:38.000000Z'
          description: date entry was last modified (content, name or location changed)
        path:
          type: string
          example: '3260/3261/3262'
          description: full path of parent folder IDs for this entry up to root
        users:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              email:
                type: string
                example: 'example@gmail.com'
...
    ShareableLink:
      type: object
      properties:
        id:
          type: integer
          format: int64
        hash:
          type: string
          example: 3h8N1nxZq7WtyRmXDIosFupSl9MCrD
          description: ID for link that will be visible in url
        password:
          type: string
          example: $2y$10$BI8kz6zFeB0CA0P425eMSe7KJcm8OKjWijHTQrRP.cjqRtynMq4Sy
          description: Passwords are hashed, so this will not match the password that was entered in plain text
        user_id:
          type: integer
          format: int64
          description: ID of user that created this link
        entry_id:
          type: integer
          format: int64
          description: ID of file entry this link is attached to
        entry:
          $ref: '#/components/schemas/FileEntry'
        expires_at:
          type: string
          example: '2021-03-06T17:34:00.000000Z'
        allow_edit:
          type: boolean
          example: false
        allow_download:
          type: boolean
          example: false
    Shareable-Link-Response:
      type: object
      properties:
        status:
          type: string
          example: success
        link:
          $ref: '#/components/schemas/ShareableLink'
        folderChildren:
          description: If link is for folder, this will list immediate children of that folder
          type: array
          items:
            $ref: '#/components/schemas/FileEntry'
        errors:
          type: object
          properties:
            some_data_1:
              type: string
              example: Error message for data 1
            some_data_2:
              type: string
              example: Error message for data 2
    401-Response:
      description: Unauthenticated. Either access token is not provided or is invalid.
    403-Response:
      description: Unauthorized access. You don't have permissions required to perform this operation.
    422-Response:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: 'Reason for the error'
        errors:
          type: object
          properties:
            some_data_1:
              type: string
              example: Error message for data 1
            some_data_2:
              type: string
              example: Error message for data 2
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer